% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsPackage-package.R
\docType{package}
\name{StatsPackage-package}
\alias{StatsPackage}
\alias{StatsPackage-package}
\title{StatsPackage}
\description{
Statistical functions for introductory inference and design.
}
\author{
\strong{Maintainer}: Chris Posadas \email{christopher.v.posadas@nps.edu}

}
\keyword{internal}
