% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsPackage-package.R
\docType{package}
\name{StatsPackage-package}
\alias{StatsPackage}
\alias{StatsPackage-package}
\title{StatsPackage}
\description{
Statistical functions for introductory inference and design.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christopherposadasp-ctrl/StatsRPackage}
  \item Report bugs at \url{https://github.com/christopherposadasp-ctrl/StatsRPackage/issues}
}

}
\author{
\strong{Maintainer}: Chris Posadas \email{christopher.v.posadas@nps.edu}

}
\keyword{internal}
